@@ -27,8 +27,8 @@
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.21",
     "prettier": "^2.8.4",
-    "tailwindcss": "^3.2.7",
-    "typescript": "~4.8.4",
+    "tailwindcss": "^3.3.0",
+    "typescript": "^5.0.2",
     "vite": "^4.1.4",
     "vue-tsc": "^1.2.0"
   }
